Bug 4120 - EBChombo: use chombo methods to calculate the total
we only calculate total in serial for vcell output, but chombo has methods to compute total which supports parallel. So we decided to use chomobo methods compute the total. However, writing post processing data in parallel is not done.  What is changed is:

1. use chombo methods to compute totals.
2. in serial, post processing data is written in either vcell output or chombo output.

still use the old way to compute totals for membrane variables.

svn path=/vcell/trunk/numerics/; revision=16983


Former-commit-id: 301c5d3f384d45d93a734d9763bb79599793aa49