Fixed two problems:
(1) Fixed index bound checking in MoleculeType::getComponentStateName().
(2) Changes to BasicRxnClass::tryToAdd(). It was possible for certain
kinds of rules that the mappingSets were not updated correctly because
the head molecule matched a reactant pattern before and after a reaction
event BUT the mapping was different after the firing. In such cases, the
mappingSet was not updated properly. The logic was changed to fix this
problem. NOTE: the changes may cause some models to execute less
efficiently.

commit edde686ba592f419147f93e11f256129f9bc2b78, aug 13, 2013